Use sats and int values for assert_fee_amount
When the feerate given to assert_fee_amount results in a fee which with
fractional satoshi values, the calculated target would be rounded to the
nearest sat value. However in the real code, the fee is always rounded
down due to integer truncation. In order for this to work correctly, the
test framework should also round down in the same way. By using ints (so
the values calculated are in satoshis), we can achieve the same effect.