rpc: use options object for walletprocesspsbt
Instead of having a ton of arguments and long chain of bools, use an
options object for all of the additional parameters to
walletprocesspsbt. The old style is still acceptable, just hidden, for
backwards compatibility. Tests are updated to use the new options
object, but the old style is still tested.